requirements.txt
easydict~=1.9
mindspore-gpu~=1.5.0
numpy~=1.21.4
opencv-python~=4.5.4.58
scipy~=1.7.2
